reworked notifications

+++ src/components/notifications-list/NotificationsItem.jsx
@@ -9,6 +9,10 @@ import {
   mapDispatchToProps,
   mapStateToProps,
 } from '../../commands/notifications'
+// eslint-disable-next-line no-unused-vars
+import { Link } from 'react-router-dom'
+import { push } from "connected-react-router"
+import {store} from '../../store/store'
 
 class NotificationsItem extends React.Component {
   constructor(props) {
@@ -19,35 +23,37 @@ class NotificationsItem extends React.Component {
     }
   }
   delete = () => {
-    this.props.deleteNotification(this.props.routeId)
+    this.props.deleteNotification(this.props.id)
   }
-
+  handleClick=(event) =>{
+    let click=event.target
+    if(click.className==="oi oi-x"){
+      this.delete()
+      this.props.handleDelete()
+    }else{
+      store.dispatch(push(this.state.routeId))
+    }
+  }
   render() {
     return (
       <ListGroup.Item
         key={this.props.routeId}
         className="itemOfNotificationList"
         /* onMouseEnter={this.props.handleAction} */
+        onClick={(event)=> this.handleClick(event)}
       >
         <Row>
-          {' '}
           <Col xs="10" md="11">
-            {' '}
-            <a href={this.state.routeId} className="black">
               {this.state.text}
-            </a>
           </Col>
           <Col xs="2" md="1">
             <span
               className="oi oi-x"
-              onClick={() => {
-                this.delete()
-                this.props.handleDelete()
-              }}
             />
           </Col>
         </Row>
       </ListGroup.Item>
+
     )
   }
 }
